Don't enforce a version bump for doc-only changes
PRs that _only_ change the chart's README or other `.md` documentation files should not automatically require a version bump and new release of that chart, if there are no changes to the functionality or dependencies.  This will allow us to add notes to a chart's README as it appears on GitHub without forcing the chart to be re-deployed to the GATE repository.