docs: bitnami postgresql image is moving
As of August 2025 the bitnami/postgresql image on which the postgresql subchart depends will no longer be available at the location that the chart expects.  It has been archived to a different path within Docker Hub, so the chart will still work but requires an additional value setting to point to the moved image location.